Move AppFileMetadata to internal package.
